Projects/QuantumSandbox/backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
qiskit==0.34.2
qiskit-aer==0.10.4
numpy==1.21.6
scipy==1.7.3
matplotlib==3.5.2
pyodide-build==0.1.0
gunicorn==20.1.0
python-dotenv==0.20.0
pytest==7.0.1
